Continued with offline Docker image handling
Added image import functionality. The twimg helper can now extract
Docker images from tar files. The tar archive needs an index.txt file in
the following format:

	file_name;repository;tag

where file_name references an image  file inside the archive, repository
the repository name of that file and tag the tag of the image.

The images need to be extracted with

	docker save repository:tag -o filename